Give the ResourceManager the real update rate
ResourceStorage sizes every joint limiter's per-cycle allowance from
ResourceManagerParams::update_rate, which defaults to 100 Hz.  This plugin
never set it, because it builds the resource manager before the
controller_manager that would normally supply it - so a 200 Hz loop enforced
every position, velocity and effort limit at twice its real allowance, with
nothing in the log to say so except a "desired period" that did not match the
loop.  Read it off our own node instead, and check it against the
controller_manager's once that exists: a mismatch now prints the ratio rather
than hiding it.

+++ mujoco_ros2_control/src/mujoco_ros2_control.cpp
@@ -216,9 +216,9 @@ void MujocoRos2Control::init()
       "update_rate mismatch: the controller_manager loop runs at %ld Hz, but the joint "
       "limiters were built for %u Hz, so every command limit is enforced at %.2fx its real "
       "value.  Set 'update_rate' on the %s node to %ld as well.",
-      update_rate, params.update_rate,
+      static_cast<long>(update_rate), params.update_rate,
       static_cast<double>(update_rate) / static_cast<double>(params.update_rate),
-      node_->get_name(), update_rate);
+      node_->get_name(), static_cast<long>(update_rate));
   }
   control_period_ = rclcpp::Duration(std::chrono::duration_cast<std::chrono::nanoseconds>(
     std::chrono::duration<double>(1.0 / static_cast<double>(update_rate))));
